requirements.txt
# Core Dependencies
streamlit>=1.50.0
nltk>=3.9.2
pdfplumber>=0.11.7
pypdf2>=3.0.1
python-docx>=1.2.0
reportlab>=4.4.4
scikit-learn>=1.7.2
spacy>=3.8.7

# ML/AI Dependencies (Required for summarization)
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Audio Processing (Optional - for audio transcription)
# Install at least one: whisper OR whisperx
openai-whisper>=20231117
# whisperx>=3.1.1  # Alternative to whisper, uncomment if preferred

# Speaker Diarization (Optional - for better speaker identification)
pyannote.audio>=3.1.0
pyannote.core>=5.0.0

# Additional dependencies for transformers
accelerate>=0.20.0
huggingface-hub>=0.16.0

# Sentence Transformers (Optional - for embeddings)
sentence-transformers>=2.2.0

# Additional utilities
numpy>=1.24.0
pandas>=2.0.0

